Pass `extraData` to `stickyOverrideRowRenderer` to re-render sticky headers
In our usage of FlashList we have a "selected" checkbox on the sticky header of each group. 

When these rows are stuck to the top, they don't update when `extraData` changes. When they come unstuck from the top, they update immediately. Regular header rows are updated when `extraData` changes, but these for some reason aren't.

We've been testing this locally and it works nicely. I don't know whether there's a specific reason that `extraData` is not included in sticky row renderers, but it seems reasonable that rows should behave the same way when they're stuck as they do when they're not stuck.